requirements.txt
Babel==2.14.0
blinker==1.7.0
cachelib==0.10.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Admin==1.6.1
Flask-BabelEx==0.9.4
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security-Too==5.3.3
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
passlib==1.7.4
requests==2.31.0
speaklater==1.3
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
uWSGI==2.0.23
Werkzeug==3.0.1
WTForms==3.1.2
